requirements.txt
# Python dependencies for ICS Security Monitoring System

# MODBUS protocol
pymodbus>=2.5.3

# Network scanning and discovery
python-nmap>=0.7.1
scapy>=2.4.5

# ML/AI libraries
numpy>=1.20.3
pandas>=1.3.4
scikit-learn>=1.0.1
tensorflow>=2.8.0
joblib>=1.1.0

# Backend
fastapi>=0.70.0
uvicorn>=0.15.0
sqlalchemy>=1.4.27
psycopg2-binary>=2.9.2
redis>=4.0.2
celery>=5.2.0
alembic>=1.7.5
pydantic>=1.8.2

# Web Sockets
websockets>=10.1

# Utilities
python-dotenv>=0.19.2
requests>=2.26.0
PyYAML>=6.0 